parasut and search

+++ database/migrations/2021_03_11_144255_invoicerecords.php
@@ -24,7 +24,7 @@ class Invoicerecords extends Migration
             $table->string('email', 255)->nullable();
             $table->string('country', 255)->nullable();
             $table->string('city', 255)->nullable();
-            $table->integer('phone')->unsigned();
+            $table->bigInteger('phone')->unsigned();
             $table->string('district', 255)->nullable();
             $table->enum('invoice_type', ['individual', 'institutional'])->default('individual');
             $table->enum('taxpayer',['false','true'])->default('false');
